requirements.txt
PyOpenGL
numpy
pygame